+++ pkg/awsvpc/vpc.go
@@ -11,68 +11,8 @@ import (
 
 	"github.com/np-guard/models/pkg/connection"
 	"github.com/np-guard/vpc-network-config-analyzer/pkg/commonvpc"
-	"github.com/np-guard/vpc-network-config-analyzer/pkg/drawio"
 	"github.com/np-guard/vpc-network-config-analyzer/pkg/vpcmodel"
 )
-
-func (rc *AWSresourcesContainer) getIgwConfig(
-	res *vpcmodel.MultipleVPCConfigs,
-	regionToStructMap map[string]*commonvpc.Region,
-	// igwToSubnet map[string][]*commonvpc.Subnet,
-	skipByVPC map[string]bool,
-) error {
-	for _, igw := range rc.InternetGWList {
-		igwId := *igw.InternetGatewayId
-		subnets := []*commonvpc.Subnet{}
-		for _, att := range igw.Attachments {
-			vpcUID := *att.VpcId
-			if skipByVPC[vpcUID] {
-				continue
-			}
-			vpc := res.GetVPC(vpcUID).(*commonvpc.VPC)
-			subnets = append(subnets, vpc.Subnets()...)
-		}
-		routerIgw := newIGW(igwId, igwId, subnets, defaultRegionName, regionToStructMap)
-		// TODO - where to put this resource?
-		for _, att := range igw.Attachments {
-			vpcUID := *att.VpcId
-			if skipByVPC[vpcUID] {
-				continue
-			}
-			res.Config(vpcUID).RoutingResources = append(res.Config(vpcUID).RoutingResources, routerIgw)
-			res.Config(vpcUID).UIDToResource[routerIgw.ResourceUID] = routerIgw
-		}
-	}
-
-	return nil
-}
-
-///////////////////////////////////////////////////////////////////////////////////////////////////
-
-func getSubnetsNodes(subnets []*commonvpc.Subnet) []vpcmodel.Node {
-	res := []vpcmodel.Node{}
-	for _, s := range subnets {
-		res = append(res, s.Nodes()...)
-	}
-	return res
-}
-
-func newIGW(igwName, igwCRN string, subnets []*commonvpc.Subnet, region string, regionToStructMap map[string]*commonvpc.Region) *InternetGateway {
-	srcNodes := getSubnetsNodes(subnets)
-	return &InternetGateway{
-		VPCResource: vpcmodel.VPCResource{
-			ResourceName: igwName,
-			ResourceUID:  igwCRN,
-			ResourceType: commonvpc.ResourceTypePublicGateway,
-			Region:       region,
-		},
-		src:        srcNodes,
-		srcSubnets: subnets,
-		region:     commonvpc.GetRegionByName(region, regionToStructMap),
-	}
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 type InternetGateway struct {
 	vpcmodel.VPCResource
@@ -100,8 +40,11 @@ func (igw *InternetGateway) ExternalIP() string {
 }
 
 func (igw *InternetGateway) AllowedConnectivity(src, dst vpcmodel.VPCResourceIntf) (*connection.Set, error) {
-	if areNodes, src1, dst1 := isNodesPair(src, dst); areNodes {
-		if vpcmodel.HasNode(igw.Sources(), src1) && dst1.IsExternal() {
+	if areNodes, srcNode, dstNode := isNodesPair(src, dst); areNodes {
+		if vpcmodel.HasNode(igw.Sources(), srcNode) && dstNode.IsExternal() {
+			return connection.All(), nil
+		}
+		if vpcmodel.HasNode(igw.Sources(), dstNode) && srcNode.IsExternal() {
 			return connection.All(), nil
 		}
 		return connection.None(), nil
@@ -112,8 +55,14 @@ func (igw *InternetGateway) AllowedConnectivity(src, dst vpcmodel.VPCResourceInt
 			if dstNode.IsExternal() && hasSubnet(igw.srcSubnets, srcSubnet) {
 				return connection.All(), nil
 			}
-			return connection.None(), nil
 		}
+		dstSubnet := src.(*commonvpc.Subnet)
+		if srcNode, ok := dst.(vpcmodel.Node); ok {
+			if srcNode.IsExternal() && hasSubnet(igw.srcSubnets, dstSubnet) {
+				return connection.All(), nil
+			}
+		}
+		return connection.None(), nil
 	}
 	return nil, errors.New("unexpected src/dst input types")
 }
@@ -135,6 +84,8 @@ func (igw *InternetGateway) StringOfRouterRules(listRulesInFilter []vpcmodel.Rul
 	return "", nil
 }
 
+// ////////////////////////////////////
+// todo - these two methods are duplicated from ibm/vpc.go needs to be reunion
 func isNodesPair(src, dst vpcmodel.VPCResourceIntf) (res bool, srcNode, dstNode vpcmodel.Node) {
 	srcNode, isSrcNode := src.(vpcmodel.Node)
 	dstNode, isDstNode := dst.(vpcmodel.Node)
@@ -150,7 +101,4 @@ func hasSubnet(listSubnets []*commonvpc.Subnet, subnet *commonvpc.Subnet) bool {
 	return false
 }
 
-func (igw *InternetGateway) GenerateDrawioTreeNode(gen *vpcmodel.DrawioGenerator) drawio.TreeNodeInterface {
-	return drawio.NewTransitGatewayTreeNode(gen.TreeNode(igw.Region()).(*drawio.RegionTreeNode), igw.Name())
-}
-func (igw *InternetGateway) ShowOnSubnetMode() bool { return true }
+/////////////////////////////////////////////////////////////
